pyproject.toml
[project]
name = "books-gen"
version = "0.1.0"
description = "Generador de libros usando LLMs, LangGraph y Groq"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-groq",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "typer",
    "rich",
    "ipykernel>=6.29.5",
    "grandalf>=0.8",
    "pydantic-settings>=2.9.1",
    "loguru>=0.7.3",
    "jinja2>=3.1.6",
    "langgraph-cli[inmem]>=0.2.10",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["books_gen"]


